barman: Create a PoC of a CloudInterface class for Storj
Create a proof of concept of creating a CloudInterface class for Storj
decentralized object storage.

The new class doesn't implement all the methods than the original
CloudInterface only the ones used by the
barman/client/cloud_walarlchive.py tool because this is a proof of
concept to see how we could adapt the barman client tools just
implementing a new CloudIterface class.