barman/client: Create cloud walarchive tool for Storj
Create a new client cloud walarchive tool for using Storj network.

This is a copy of the orignal barman/clients/cloud_walarchive.py file
with the minimal modifications for using the StorjCloudInterface class
and verify that the proof of concept works.

In a proper implementation the tools should be the same and have some
command line arguments to specify which cloud provider to use (Storj or
S3).